DAH-1123 Move aside component to appropriate module, add isApplicationOpen prop
