+++ app/javascript/pages/listings/listing-detail.tsx
@@ -13,7 +13,7 @@ import { getListing } from "../../api/listingApiService"
 import { RailsListing } from "../../modules/listings/SharedHelpers"
 import { ListingDetailsImageCard } from "../../modules/listingDetails/ListingDetailsImageCard"
 import { ListingDetailsPricingTable } from "../../modules/listingDetails/ListingDetailsPricingTable"
-import { ListingDetailsAside } from "../../modules/listingDetails/ListingDetailsAside"
+import { ListingDetailsAside } from "../../modules/listingDetailsAside/ListingDetailsAside"
 import { ListingDetailsEligibility } from "../../modules/listingDetails/ListingDetailsEligibility"
 import { ListingDetailsFeatures } from "../../modules/listingDetails/ListingDetailsFeatures"
 import { ListingDetailsNeighborhood } from "../../modules/listingDetails/ListingDetailsNeighborhood"
